--- src/main/java/es/in2/blockchain/connector/core/service/TransactionService.java
@@ -1,9 +1,0 @@
-package es.in2.blockchain.connector.core.service;
-
-import es.in2.blockchain.connector.core.domain.Transaction;
-
-
-public interface TransactionService {
-    Transaction createTransaction(String entityId, String entityHash, String datalocation);
-    Transaction editTransaction(Transaction transaction);
-}
