+++ src/test/java/es/in2/blockchain/connector/core/service/OnChainEntityServiceImplTest.java
@@ -1,6 +1,7 @@
 package es.in2.blockchain.connector.core.service;
 
 import es.in2.blockchain.connector.core.exception.RequestErrorException;
+import es.in2.blockchain.connector.core.repository.TransactionRepository;
 import es.in2.blockchain.connector.core.service.impl.OnChainEntityServiceImpl;
 import es.in2.blockchain.connector.integration.blockchainnode.configuration.BlockchainNodeIConfig;
 import es.in2.blockchain.connector.integration.blockchainnode.configuration.BlockchainNodeProperties;
@@ -31,7 +32,7 @@ import static org.mockito.Mockito.when;
     private BlockchainNodeProperties blockchainNodeProperties;
 
     @Mock
-    private TransactionService transactionService;
+    private TransactionRepository transactionRepository;
 
     @InjectMocks
     private OnChainEntityServiceImpl onChainEntityService;
@@ -39,7 +40,7 @@ import static org.mockito.Mockito.when;
     @BeforeEach
     public void setUp() {
         MockitoAnnotations.openMocks(this);
-        onChainEntityService = new OnChainEntityServiceImpl(hashLinkService, blockchainNodeIConfig, blockchainNodeProperties, transactionService);
+        onChainEntityService = new OnChainEntityServiceImpl(hashLinkService, blockchainNodeIConfig, blockchainNodeProperties, transactionRepository);
     }
 
     @Test
